ajout de la fonction update

+++ src/Controller/PhoneController.php
@@ -3,8 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\Phone;
+use App\Form\PhoneType;
 use App\Repository\PhoneRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\Form\FormInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -24,7 +26,6 @@ class PhoneController extends AbstractController
     public function showAction(Phone $phone, PhoneRepository $phoneRepository, Request $request)
     {
         
-        //$phone = $phoneRepository->find($phone->getId());
         return $this->json($phone, 200, [], ['groups' => 'phone:read']);
     }
     
@@ -82,17 +83,71 @@ class PhoneController extends AbstractController
         
     }
 
+    /**
+     * @Route("/api/phones/{id}", name="app_phones_update", methods={"PUT"})
+     */
+    public function updateAction(Phone $phone,PhoneRepository $phoneRepository, Request $request, SerializerInterface $serializer, EntityManagerInterface $em, ValidatorInterface $validator)
+    {
+        if (!$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            return $this->json([
+                'status' => 401,
+                'message' => "Acces Denied"
+            ], 401);
+        }
+        
+        try{
+        
+        $jsonRecu = $request->getContent();
+        $phoneValidate = $serializer->deserialize($jsonRecu, Phone::class, 'json');
+        
+        $errors = $validator->validate($phoneValidate);
+        
+        if(count($errors) > 0) {
+            return $this->json($errors, 400);
+        }
+        
+        $form = $this->createForm(PhoneType::class, $phone);
+        $this->processForm($request, $form);
+        
+        }catch(NotEncodableValueException $e){
+            return $this->json([
+                'status' => 400,
+                'message' => $e->getMessage()
+            ], 400);
+        }
+
+        $em->persist($phone);
+        $em->flush();
+
+        return $this->json([
+            'status' => 200,
+            'message' => "le téléphone a bien été mis à jour"
+        ], 200);
+        
+    }
+
      /**
      * @Route("/api/phones/{id}", name="app_phones_delete", methods={"DELETE"})
      */
     public function deleteAction(Phone $phone, Request $request, EntityManagerInterface $em)
     {
-        //$phone = $phoneRepository->find($phone->getId());
-
+        if (!$this->get('security.authorization_checker')->isGranted('ROLE_ADMIN')) {
+            return $this->json([
+                'status' => 401,
+                'message' => "Acces Denied"
+            ], 401);
+        }
+        
         $em->remove($phone);
         $em->flush();
         
         return $this->json(null, 204);
+    }
+
+    private function processForm(Request $request, FormInterface $form)
+    {
+        $data = json_decode($request->getContent(), true);
+        $form->submit($data);
     }
 
 }
